requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.1
django-bootstrap-v5==1.0.11
django-markdownx==4.0.2
django-non-dark-admin==2.0.2
Pillow==9.5.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
requests~=2.31.0